+++ packages/build-utils/src/types.ts
@@ -610,7 +610,6 @@ export interface Service {
   handlerFunction?: string;
   /* worker/job service config */
   topics?: ServiceTopics;
-  consumer?: string;
   /** custom prefix to inject service URL env vars */
   envPrefix?: string;
 }
@@ -898,7 +897,6 @@ export interface ExperimentalServiceConfig {
 
   /* Worker/job service config */
   topics?: ServiceTopics;
-  consumer?: string;
 
   /** Custom prefix to use to inject service URL env vars */
   envPrefix?: string;