+++ packages/config/src/types.ts
@@ -659,10 +659,6 @@ export interface VercelConfig {
        */
       topics?: string[] | ServiceQueueTopic[];
       /**
-       * Consumer group name for worker subscription.
-       */
-      consumer?: string;
-      /**
        * Custom prefix to use to inject service URL env vars.
        */
       envPrefix?: string;