+++ packages/koa/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vercel/koa
 
+## 0.1.52
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.7.11
+
 ## 0.1.51
 
 ### Patch Changes